Split the codebase for multi-protocol support
