+++ src/util.rs
@@ -1,0 +1,6 @@
+use std::process::exit;
+
+pub fn fatal(msg: &str) -> ! {
+    eprintln!("FATAL: {}", msg);
+    exit(1);
+}
